回答音声の再生が終わったら次の音声入力に移行できるように修正
