+++ src/App.tsx
@@ -24,32 +24,38 @@ function App() {
     ['user', 'こんにちは！あなたはカレー作りの名人です。それを踏まえてこれからの回答に答えてください。'],
   ]);
 
-  const handleUploadResult = (speechScript: string, speechBase64: string, answer: string) => {
+  const handleUploadResult = (speechScript: string, answer: string) => {
     // chatAPIの回答をchatに追加
     if(speechScript) setChat((prevChat) => [...prevChat, ['assistant', speechScript]]);
     if(answer) setChat((prevChat) => [...prevChat, ['assistant', answer]]);
     // Base64エンコードされたMP3を再生する処理
-    playAudioFromBase64(speechBase64);
+    // playAudioFromBase64(speechBase64);
   };
 
   // Base64エンコードされたMP3を再生する関数
-  const playAudioFromBase64 = (base64Audio: string) => {
-    // Base64文字列からバイナリデータに変換
-    const binaryString = window.atob(base64Audio);
-    const len = binaryString.length;
-    const bytes = new Uint8Array(len);
+  // const playAudioFromBase64 = (base64Audio: string) => {
+  //   // Base64文字列からバイナリデータに変換
+  //   const binaryString = window.atob(base64Audio);
+  //   const len = binaryString.length;
+  //   const bytes = new Uint8Array(len);
 
-    for (let i = 0; i < len; i++) {
-      bytes[i] = binaryString.charCodeAt(i);
-    }
+  //   for (let i = 0; i < len; i++) {
+  //     bytes[i] = binaryString.charCodeAt(i);
+  //   }
 
-    const blob = new Blob([bytes], { type: 'audio/mp3' });
-    const url = URL.createObjectURL(blob);
+  //   const blob = new Blob([bytes], { type: 'audio/mp3' });
+  //   const url = URL.createObjectURL(blob);
 
-    // Audioオブジェクトを使って音声を再生
-    const audio = new Audio(url);
-    audio.play().catch((error) => console.error('音声の再生に失敗しました:', error));
-  };
+  //   // Audioオブジェクトを使って音声を再生
+  //   const audio = new Audio(url);
+  //   audio.play().catch((error) => console.error('音声の再生に失敗しました:', error));
+
+  //   // 音声の再生が終了したら次の音声入力を開始
+  //   audio.onended = () => {
+  //     console.log('音声の再生が終了しました');
+  //     document.querySelector('button')?.click(); // 次の音声入力を開始
+  //   };
+  // };
 
   return (
     <Box>
